산소 view 수정완료 & 산소 평균 수정필요

+++ app/src/main/java/com/coai/samin_total/Oxygen/SetOxygenViewData.kt
@@ -9,7 +9,7 @@ data class SetOxygenViewData(
     var isAlert: Boolean = false,
     var setValue: Float = 0f,
     var setMinValue: Float = 20f,
-    var setMaxValue: Int = 23,
+    var setMaxValue: Float = 23f,
     val modelByte: Byte = 4,
     var heartbeatCount : UByte = 0u
